+++ oshi-core/src/main/java/oshi/hardware/CentralProcessor.java
@@ -33,6 +33,7 @@ import java.util.regex.Pattern;
 
 import oshi.annotation.concurrent.Immutable;
 import oshi.annotation.concurrent.ThreadSafe;
+import oshi.software.os.OSProcess;
 import oshi.software.os.OSThread;
 import oshi.util.Constants;
 import oshi.util.FileUtil;
@@ -225,10 +226,11 @@ public interface CentralProcessor {
     /**
      * Get the number of system-wide context switches which have occurred.
      * <p>
-     * Not available system-wide on macOS. Thread-level context switches are
-     * available from {@link OSThread#getContextSwitches()}.
+     * Not available system-wide on macOS. Process- and Thread-level context
+     * switches are available from {@link OSProcess#getContextSwitches()} and
+     * {@link OSThread#getContextSwitches()}.
      *
-     * @return The number of context switches, if this information is available; -1
+     * @return The number of context switches, if this information is available; 0
      *         otherwise.
      */
     long getContextSwitches();
@@ -238,7 +240,7 @@ public interface CentralProcessor {
      * <p>
      * Not available system-wide on macOS.
      *
-     * @return The number of interrupts, if this information is available; -1
+     * @return The number of interrupts, if this information is available; 0
      *         otherwise.
      */
     long getInterrupts();