+++ oshi-core/src/main/java/oshi/hardware/platform/unix/solaris/SolarisCentralProcessor.java
@@ -242,7 +242,7 @@ final class SolarisCentralProcessor extends AbstractCentralProcessor {
         for (String s : kstat) {
             swtch += ParseUtil.parseLastLong(s, 0L);
         }
-        return swtch > 0 ? swtch : -1L;
+        return swtch;
     }
 
     @Override
@@ -252,6 +252,6 @@ final class SolarisCentralProcessor extends AbstractCentralProcessor {
         for (String s : kstat) {
             intr += ParseUtil.parseLastLong(s, 0L);
         }
-        return intr > 0 ? intr : -1L;
+        return intr;
     }
 }