+++ oshi-core/src/main/java/oshi/software/common/AbstractOSProcess.java
@@ -73,16 +73,6 @@ public abstract class AbstractOSProcess implements OSProcess {
     }
 
     @Override
-    public long getMinorFaults() {
-        return 0L;
-    }
-
-    @Override
-    public long getMajorFaults() {
-        return 0L;
-    }
-
-    @Override
     public String toString() {
         StringBuilder builder = new StringBuilder("OSProcess@");
         builder.append(Integer.toHexString(hashCode()));