+++ oshi-core/src/test/java/oshi/hardware/CentralProcessorTest.java
@@ -40,8 +40,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.api.TestInstance.Lifecycle;
 
-import com.sun.jna.Platform;
-
 import oshi.SystemInfo;
 import oshi.hardware.CentralProcessor.ProcessorIdentifier;
 import oshi.hardware.CentralProcessor.TickType;
@@ -133,17 +131,10 @@ class CentralProcessorTest {
                 p.getPhysicalProcessorCount(), is(greaterThanOrEqualTo(p.getPhysicalPackageCount())));
         assertThat("Central Processor's physical package count should be higher than 0", p.getPhysicalPackageCount(),
                 is(greaterThan(0)));
-        if (Platform.isMac()) {
-            // these are unavailable on macOS
-            assertThat("Central Processor's context switch count should be -1 on macOS", p.getContextSwitches(),
-                    is(-1L));
-            assertThat("Central Processor's interrupt count should be -1 on macOS", p.getInterrupts(), is(-1L));
-        } else {
-            assertThat("Central Processor's context switch count should be 0 or higher", p.getContextSwitches(),
-                    is(greaterThanOrEqualTo(0L)));
-            assertThat("Central Processor's interrupt count should be 0 or higher", p.getInterrupts(),
-                    is(greaterThanOrEqualTo(0L)));
-        }
+        assertThat("Central Processor's context switch count should be 0 or higher", p.getContextSwitches(),
+                is(greaterThanOrEqualTo(0L)));
+        assertThat("Central Processor's interrupt count should be 0 or higher", p.getInterrupts(),
+                is(greaterThanOrEqualTo(0L)));
         for (int lp = 0; lp < p.getLogicalProcessorCount(); lp++) {
             assertThat("Logical processor number is negative", p.getLogicalProcessors().get(lp).getProcessorNumber(),
                     is(greaterThanOrEqualTo(0)));