+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # 5.6.2 (in progress)
 
+##### New Features
+* [#1587](https://github.com/oshi/oshi/pull/1587): Add context switches to OSProcess - [@dbwiddis](https://github.com/dbwiddis).
+
 ##### Bug fixes / Improvements
 * [#1585](https://github.com/oshi/oshi/pull/1585): macOS doesn't provide system level context switches or interrupts - [@dbwiddis](https://github.com/dbwiddis).